pyproject.toml
[project]
name = "supernovae"
version = "0.1.0"
description = "Supernovae dataset"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0",
    "djangorestframework>=3.16.1",
    "drf-yasg>=1.21.11",
    "factory-boy>=3.3.1",
]
